pyproject.toml
[project]
name = "vspe-gpt"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "murasame", email = "murasame.usa@gmail.com" }
]
dependencies = [
    "flask>=3.1.0",
    "openai>=1.57.2",
    "flask-cors>=5.0.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/vspe_gpt"]
